# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Response:
    properties:
      data: {}
      status:
        $ref: '#/definitions/model.status'
    type: object
  model.status:
    enum:
    - success
    - error
    type: string
    x-enum-varnames:
    - success
    - err
  pagination.Pagination:
    properties:
      current_page:
        example: 1
        type: integer
      per_page:
        example: 1
        type: integer
      total:
        example: 10
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  post.PostResponse:
    properties:
      age:
        type: integer
      animal_type:
        type: string
      author_username:
        type: string
      color:
        type: string
      comments:
        type: integer
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      gender:
        type: string
      is_favourite:
        type: boolean
      photo:
        items:
          type: integer
        type: array
      status:
        type: string
      title:
        type: string
    type: object
  post.Response:
    properties:
      meta:
        $ref: '#/definitions/pagination.Pagination'
      posts:
        items:
          $ref: '#/definitions/post.PostResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the SOS Kotopes project. It provides endpoints
    for managing the database of animals in need of help.
  title: SOS Kotopes API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get all posts
      operationId: get-all-posts
      parameters:
      - description: Limit
        in: query
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: Animal type
        in: query
        name: animal_type
        type: string
      - description: Gender
        in: query
        name: gender
        type: string
      - description: Color
        in: query
        name: color
        type: string
      - description: Location
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/post.Response'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all posts
      tags:
      - post
    post:
      consumes:
      - application/json
      description: Create a post
      operationId: create-post
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Content
        in: formData
        name: content
        required: true
        type: string
      - description: Animal type
        in: formData
        name: animal_type
        required: true
        type: string
      - description: Photo
        in: formData
        name: photo
        required: true
        type: file
      - description: Age
        in: formData
        name: age
        required: true
        type: integer
      - description: Color
        in: formData
        name: color
        required: true
        type: string
      - description: Gender
        in: formData
        name: gender
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      - description: Status
        in: formData
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/post.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Create a post
      tags:
      - post
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post
      operationId: delete-post
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Delete a post
      tags:
      - post
    get:
      consumes:
      - application/json
      description: Get post by ID
      operationId: get-post-by-id
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/post.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get post by ID
      tags:
      - post
    patch:
      consumes:
      - application/json
      description: Update a post
      operationId: update-post
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Title
        in: formData
        name: title
        type: string
      - description: Content
        in: formData
        name: content
        type: string
      - description: Animal type
        in: formData
        name: animal_type
        type: string
      - description: Photo
        in: formData
        name: photo
        type: file
      - description: Age
        in: formData
        name: age
        type: integer
      - description: Color
        in: formData
        name: color
        type: string
      - description: Gender
        in: formData
        name: gender
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Status
        in: formData
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/post.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Update a post
      tags:
      - post
...
  /user/{id}/posts:
    get:
      consumes:
      - application/json
      description: Get posts by user ID
      operationId: get-posts-user-by-id
      parameters:
      - description: User ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Limit
        in: query
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: Animal type
        in: query
        name: animal_type
        type: string
      - description: Gender
        in: query
        name: gender
        type: string
      - description: Color
        in: query
        name: color
        type: string
      - description: Location
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/post.Response'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get posts by user ID
      tags:
      - post
...
securityDefinitions:
  ApiKeyAuthBasic:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"